refactor(dummies): add file extension on import paths
