+++ src/dummies/DummyIrcClient.ts
@@ -1,8 +1,9 @@
-import { IIrcClient } from '..';
-import * as irc from '../libs/irc';
-import { parser, MpCommand, StatResult, StatStatuses } from '../parsers';
-import { escapeUserName } from '../Player';
-import { MpSettingsCase } from '../tests/cases/MpSettingsCases';
+import { IIrcClient } from '../IIrcClient.js';
+import * as irc from '../libs/irc.js';
+import { parser, MpCommand } from '../parsers/CommandParser.js';
+import { StatResult, StatStatuses } from '../parsers/StatParser.js';
+import { escapeUserName } from '../Player.js';
+import { MpSettingsCase } from '../tests/cases/MpSettingsCases.js';
 import log4js from 'log4js';
 import { EventEmitter } from 'events';
 const logger = log4js.getLogger("irc");
